requirements.txt
acme==0.39.0
aiohttp==3.6.1
aiohttp-cors==0.7.0
appdirs==1.4.3
astral==1.10.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
boto3==1.10.34
botocore==1.13.34
btest==0.59
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
click-datetime==0.2
configparser==4.0.2
construct==2.9.45
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
distro==1.4.0
dnspython==1.16.0
docutils==0.15.2
dpkt==1.9.2
ecdsa==0.14.1
entrypoints==0.3
envs==1.3
future==0.18.2
gitdb2==2.0.6
GitPython==3.0.5
graphviz==0.13.2
hass-nabucasa==0.29
home-assistant-frontend==20191119.6
homeassistant==0.102.3
idna==2.8
ifaddr==0.1.6
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
josepy==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
maxminddb==1.5.2
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.2
multidict==4.6.1
nbconvert==5.6.1
nbformat==4.4.0
netdisco==2.6.0
netifaces==0.10.9
networkx==2.4
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
plotly==4.4.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.9.4
pygeoip==0.3.2
Pygments==2.5.2
pygraphviz==1.5
pyHS100==0.3.5
PyJWT==1.7.1
PyMetno==0.4.6
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.5
PyQRCode==1.2.1
PyQt5==5.14.0
PyQt5-sip==12.7.0
pyRFC3339==1.1
pyrsistent==0.15.6
python-dateutil==2.8.0
python-jose-cryptodome==1.3.2
python-miio==0.4.7
python-slugify==4.0.0
pytz==2019.3
PyXiaomiGateway==0.12.4
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-toolbelt==0.9.1
retrying==1.3.3
ruamel.yaml==0.15.100
s3transfer==0.2.1
scapy==2.4.3
scapy-http==1.8.2
scipy==1.4.1
seaborn==0.9.0
semantic-version==2.8.4
Send2Trash==1.5.0
six==1.13.0
smmap2==2.0.5
snitun==0.20
SQLAlchemy==1.3.10
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tk==0.1.0
tornado==6.0.3
tqdm==4.40.1
traitlets==4.3.3
typing==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.7
voluptuous==0.11.7
voluptuous-serialize==2.3.0
warrant==0.6.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
xmltodict==0.12.0
yarl==1.4.2
zeroconf==0.23.0
zipp==0.6.0
zkg==2.0.7
